ai102-promptflow/requirements.txt

azure-cognitiveservices-speech==1.37.0
azure-storage-blob==12.22.0
azure-identity==1.17.1
